# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Football Recruitment Intelligence",
    "description": "Versioned aggregate research contracts and precomputed player similarity; no arbitrary SQL.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/matches": {
      "get": {
        "summary": "Matches",
        "operationId": "matches_api_v1_matches_get",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "statsbomb",
                    "skillcorner"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchSummary"
                  },
                  "title": "Response Matches Api V1 Matches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explorer/{match_id}": {
      "get": {
        "summary": "Explorer",
        "operationId": "explorer_api_v1_explorer__match_id__get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Explorer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/players": {
      "get": {
        "summary": "Players",
        "operationId": "players_api_v1_players_get",
        "parameters": [
          {
            "name": "match_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerSummary"
                  },
                  "title": "Response Players Api V1 Players Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/player-dna/{player_id}": {
      "get": {
        "summary": "Dna Profile",
        "operationId": "dna_profile_api_v1_player_dna__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Player Id"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 900,
              "title": "Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DNAProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/player-dna/{player_id}/similar": {
      "get": {
        "summary": "Dna Similar",
        "operationId": "dna_similar_api_v1_player_dna__player_id__similar_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Player Id"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 900,
              "title": "Threshold"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DNANeighbor"
                  },
                  "title": "Response Dna Similar Api V1 Player Dna  Player Id  Similar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/translation/players/{player_id}": {
      "get": {
        "summary": "Translation Player",
        "operationId": "translation_player_api_v1_translation_players__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationPlayerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/translation/predict": {
      "get": {
        "summary": "Translation Predict",
        "operationId": "translation_predict_api_v1_translation_predict_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Player Id"
            }
          },
          {
            "name": "source_environment",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Environment"
            }
          },
          {
            "name": "target_environment",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Target Environment"
            }
          },
          {
            "name": "target_role",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 10,
              "title": "Target Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationPrediction"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DNAContribution": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "standardized_difference": {
            "type": "number",
            "title": "Standardized Difference"
          },
          "squared_contribution": {
            "type": "number",
            "minimum": 0.0,
            "title": "Squared Contribution"
          },
          "share": {
            "type": "number",
            "maximum": 1.000001,
            "minimum": 0.0,
            "title": "Share"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "feature",
          "standardized_difference",
          "squared_contribution",
          "share"
        ],
        "title": "DNAContribution"
      },
...
      "DNAFeatureValue": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "percentile": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "value",
          "percentile"
        ],
        "title": "DNAFeatureValue"
      },
...
      "DNANeighbor": {
        "properties": {
          "player_id": {
            "type": "string",
            "title": "Player Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "teams": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Teams"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "minutes": {
            "type": "number",
            "title": "Minutes"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "distance": {
            "type": "number",
            "minimum": 0.0,
            "title": "Distance"
          },
          "stability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Stability"
          },
          "similar": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Similar"
          },
          "different": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Different"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/DNAContribution"
            },
            "type": "array",
            "title": "Contributions"
          },
          "family_contributions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Family Contributions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "player_id",
          "name",
          "teams",
          "role",
          "minutes",
          "rank",
          "distance",
          "stability",
          "similar",
          "different",
          "contributions",
          "family_contributions"
        ],
        "title": "DNANeighbor"
      },
...
      "DNAProfile": {
        "properties": {
          "player": {
            "$ref": "#/components/schemas/DNAPlayer"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "player-dna-v1"
          },
          "cohort": {
            "type": "string",
            "title": "Cohort"
          },
          "threshold": {
            "type": "integer",
            "title": "Threshold"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          },
          "exclusions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclusions"
          },
          "comparison_size": {
            "type": "integer",
            "title": "Comparison Size"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/DNAFeatureValue"
            },
            "type": "array",
            "title": "Features"
          },
          "scaled_vector": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scaled Vector"
          },
          "neighbors": {
            "items": {
              "$ref": "#/components/schemas/DNANeighbor"
            },
            "type": "array",
            "title": "Neighbors"
          },
          "neighbor_jaccard": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighbor Jaccard"
          },
          "bootstrap_samples": {
            "type": "integer",
            "title": "Bootstrap Samples"
          },
          "bins": {
            "items": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Bins"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "player",
          "cohort",
          "threshold",
          "eligible",
          "exclusions",
          "comparison_size",
          "features",
          "scaled_vector",
          "neighbors",
          "neighbor_jaccard",
          "bootstrap_samples",
          "bins"
        ],
        "title": "DNAProfile"
      },
      "EventCount": {
        "properties": {
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_type",
          "count"
        ],
        "title": "EventCount"
      },
      "Explorer": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "match": {
            "$ref": "#/components/schemas/MatchSummary"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerSummary"
            },
            "type": "array",
            "title": "Players"
          },
          "event_counts": {
            "items": {
              "$ref": "#/components/schemas/EventCount"
            },
            "type": "array",
            "title": "Event Counts"
          },
          "spatial_bins": {
            "items": {
              "$ref": "#/components/schemas/SpatialBin"
            },
            "type": "array",
            "title": "Spatial Bins"
          },
          "spatial_sample_size": {
            "type": "integer",
            "title": "Spatial Sample Size"
          },
          "tracking_snapshots": {
            "items": {
              "$ref": "#/components/schemas/TrackingSnapshot"
            },
            "type": "array",
            "title": "Tracking Snapshots"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "match",
          "players",
          "event_counts",
          "spatial_bins",
          "spatial_sample_size",
          "tracking_snapshots"
        ],
        "title": "Explorer"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "competition": {
            "type": "string",
            "title": "Competition"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "home": {
            "type": "string",
            "title": "Home"
          },
          "away": {
            "type": "string",
            "title": "Away"
          },
          "score": {
            "type": "string",
            "title": "Score"
          },
          "coverage": {
            "type": "string",
            "title": "Coverage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "provider",
          "competition",
          "season",
          "date",
          "home",
          "away",
          "score",
          "coverage"
        ],
        "title": "MatchSummary"
      },
...
      "PlayerSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "match_id": {
            "type": "string",
            "title": "Match Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "team": {
            "type": "string",
            "title": "Team"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "minutes_method": {
            "type": "string",
            "title": "Minutes Method"
          },
          "shots": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shots"
          },
          "passes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passes"
          },
          "xg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xg"
          },
          "shots_per90": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shots Per90"
          },
          "pass_completion": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Completion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "match_id",
          "provider",
          "name",
          "team",
          "position",
          "minutes",
          "minutes_method",
          "shots",
          "passes",
          "xg",
          "shots_per90",
          "pass_completion"
        ],
        "title": "PlayerSummary"
      },
...
      "SpatialBin": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y",
          "count"
        ],
        "title": "SpatialBin"
      },
...
      "TrackingSnapshot": {
        "properties": {
          "frame": {
            "type": "integer",
            "title": "Frame"
          },
          "timestamp_seconds": {
            "type": "number",
            "title": "Timestamp Seconds"
          },
          "period": {
            "type": "integer",
            "title": "Period"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackingPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "frame",
          "timestamp_seconds",
          "period",
          "points"
        ],
        "title": "TrackingSnapshot"
      },
...
      "TranslationPlayerDetail": {
        "properties": {
          "player": {
            "$ref": "#/components/schemas/TranslationPlayer"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "environments": {
            "items": {
              "$ref": "#/components/schemas/TranslationEnvironment"
            },
            "type": "array",
            "title": "Environments"
          },
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/TranslationPrediction"
            },
            "type": "array",
            "title": "Predictions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "player",
          "model_version",
          "environments",
          "predictions"
        ],
        "title": "TranslationPlayerDetail"
      },
      "TranslationPrediction": {
        "properties": {
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "source_environment_id": {
            "type": "string",
            "title": "Source Environment Id"
          },
          "target_environment_id": {
            "type": "string",
            "title": "Target Environment Id"
          },
          "target_role": {
            "type": "string",
            "title": "Target Role"
          },
          "status": {
            "type": "string",
            "enum": [
              "supported",
              "insufficient_evidence",
              "out_of_scope"
            ],
            "title": "Status"
          },
          "exclusions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclusions"
          },
          "prediction_minutes": {
            "type": "integer",
            "title": "Prediction Minutes"
          },
          "estimates": {
            "items": {
              "$ref": "#/components/schemas/TranslationEstimate"
            },
            "type": "array",
            "title": "Estimates"
          },
          "research_estimates": {
            "items": {
              "$ref": "#/components/schemas/TranslationEstimate"
            },
            "type": "array",
            "title": "Research Estimates"
          },
          "evidence": {
            "$ref": "#/components/schemas/TranslationEvidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model_version",
          "source_environment_id",
          "target_environment_id",
          "target_role",
          "status",
          "exclusions",
          "prediction_minutes",
          "estimates",
          "research_estimates",
          "evidence"
        ],
        "title": "TranslationPrediction"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}